+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- AverageCheck to compute AVG over a column.
+- MaxCheck to compute MAX over a column.
+- MinCheck to compute MIN over a column.
+
 ## [0.1.0] - 2025-12-09
 
 ### Added